Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTx
Flask-Swagger-UI
Flask-JWT-Extended
psycopg2-binary
pycryptodome
python-dotenv
openpyxl
ezdxf
flask-cors
shapely
gunicorn
boto3
botocore